# packages/contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/inventory/stock": {
      "get": {
        "operationId": "listStock",
        "parameters": [
          {
            "name": "warehouseId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "lowStock",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockLevelListDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Current stock levels across warehouses.",
        "tags": [
          "inventory"
        ]
      }
    },
    "/inventory/movements": {
      "get": {
        "operationId": "listStockMovements",
        "parameters": [
          {
            "name": "warehouseId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockMovementListDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "The stock movement ledger.",
        "tags": [
          "inventory"
        ]
      }
    },
    "/inventory/adjustments": {
      "post": {
        "operationId": "adjustStock",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustStockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockLevelDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Post a manual stock adjustment.",
        "tags": [
          "inventory"
        ]
      }
    },
    "/inventory/transfers": {
      "post": {
        "operationId": "transferStock",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferStockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockLevelDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Transfer stock between warehouses.",
        "tags": [
          "inventory"
        ]
      }
    },
    "/inventory/warehouses/{id}/stock": {
      "get": {
        "operationId": "listWarehouseStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockLevelDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Stock held in one warehouse.",
        "tags": [
          "inventory"
        ]
      }
    },
...
  },
  "info": {
    "title": "Aivoryx Platform API",
    "description": "Code-first OpenAPI contract. Regenerate with `pnpm contracts:generate`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aivoryx_session",
        "description": "HTTP-only session cookie (ADR 0010). Set by the auth endpoints in Phase 2."
      }
    },
    "schemas": {
...
      "ApiErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "example": "AIV-01J9Z8ABCDEF0123456789"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message",
          "correlationId"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetailDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "StockLevelDto": {
        "type": "object",
        "properties": {
          "warehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "warehouseCode": {
            "type": "string"
          },
          "warehouseName": {
            "type": "string"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "productSku": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "unitCode": {
            "type": "string"
          },
          "onHand": {
            "type": "string"
          },
          "reserved": {
            "type": "string"
          },
          "available": {
            "type": "string",
            "description": "onHand - reserved"
          },
          "reorderLevel": {
            "type": "string",
            "nullable": true
          },
          "lowStock": {
            "type": "boolean",
            "description": "onHand <= reorderLevel"
          }
        },
        "required": [
          "warehouseId",
          "warehouseCode",
          "warehouseName",
          "productId",
          "productSku",
          "productName",
          "unitCode",
          "onHand",
          "reserved",
          "available",
          "reorderLevel",
          "lowStock"
        ]
      },
      "StockLevelListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockLevelDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "StockMovementDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "warehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "warehouseCode": {
            "type": "string"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "productSku": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          },
          "onHandDelta": {
            "type": "string"
          },
          "reservedDelta": {
            "type": "string"
          },
          "projectId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "referenceType": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "warehouseId",
          "warehouseCode",
          "productId",
          "productSku",
          "quantity",
          "onHandDelta",
          "reservedDelta",
          "projectId",
          "referenceType",
          "notes",
          "createdAt"
        ]
      },
      "StockMovementListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockMovementDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "AdjustStockDto": {
        "type": "object",
        "properties": {
          "warehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "delta": {
            "type": "string",
            "description": "Signed decimal string — the on-hand delta."
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          },
          "idempotencyKey": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "warehouseId",
          "productId",
          "delta",
          "reason"
        ]
      },
      "TransferStockDto": {
        "type": "object",
        "properties": {
          "sourceWarehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "destinationWarehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "string",
            "description": "Decimal string, > 0."
          },
          "notes": {
            "type": "string",
            "maxLength": 500
          },
          "idempotencyKey": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "sourceWarehouseId",
          "destinationWarehouseId",
          "productId",
          "quantity"
        ]
      },
...
    }
  }
}